% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokenization.R
\name{.process_vocab}
\alias{.process_vocab}
\alias{.process_vocab.default}
\alias{.process_vocab.wordpiece_vocabulary}
\alias{.process_vocab.character}
\title{Process a Vocabulary for Tokenization}
\usage{
.process_vocab(v)

\method{.process_vocab}{default}(v)

\method{.process_vocab}{wordpiece_vocabulary}(v)

\method{.process_vocab}{character}(v)
}
\arguments{
\item{v}{An object of class \code{wordpiece_vocabulary} or a character vector.}
}
\value{
A character vector of tokens for tokenization.
}
\description{
Process a Vocabulary for Tokenization
}
\keyword{internal}
