% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokenization.R
\name{.process_wp_vocab}
\alias{.process_wp_vocab}
\alias{.process_wp_vocab.default}
\alias{.process_wp_vocab.wordpiece_vocabulary}
\alias{.process_wp_vocab.integer}
\alias{.process_wp_vocab.character}
\title{Process a Wordpiece Vocabulary for Tokenization}
\usage{
.process_wp_vocab(v)

\method{.process_wp_vocab}{default}(v)

\method{.process_wp_vocab}{wordpiece_vocabulary}(v)

\method{.process_wp_vocab}{integer}(v)

\method{.process_wp_vocab}{character}(v)
}
\arguments{
\item{v}{An object of class \code{wordpiece_vocabulary}.}
}
\value{
A character vector of tokens for tokenization.
}
\description{
Process a Wordpiece Vocabulary for Tokenization
}
\keyword{internal}
