% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vocab.R
\name{.validate_wordpiece_vocabulary}
\alias{.validate_wordpiece_vocabulary}
\title{Validator for Objects of Class wordpiece_vocabulary}
\usage{
.validate_wordpiece_vocabulary(vocab)
}
\arguments{
\item{vocab}{wordpiece_vocabulary object to validate}
}
\value{
\code{vocab} if the object passes the checks. Otherwise, abort with
message.
}
\description{
Validator for Objects of Class wordpiece_vocabulary
}
\keyword{internal}
