% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{op-null-default}
\alias{op-null-default}
\alias{\%||\%}
\title{Default value for \code{NULL}}
\usage{
x \%||\% y
}
\arguments{
\item{x, y}{If \code{x} is NULL, will return \code{y}; otherwise returns \code{x}.}
}
\description{
Mostly copied from rlang package.
}
\examples{
1 \%||\% 2
NULL \%||\% 2
}
