% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{\%||\%}
\alias{\%fin\%}
\alias{fmatch}
\alias{wordpiece_vocab}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{fastmatch}{\code{\link[fastmatch:fmatch]{\%fin\%}}, \code{\link[fastmatch]{fmatch}}}

  \item{rlang}{\code{\link[rlang:op-null-default]{\%||\%}}}

  \item{wordpiece.data}{\code{\link[wordpiece.data]{wordpiece_vocab}}}
}}

