% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{wordpiece_cache_dir}
\alias{wordpiece_cache_dir}
\title{Retrieve Directory for wordpiece Cache}
\usage{
wordpiece_cache_dir()
}
\value{
A character vector with the normalized path to the cache.
}
\description{
The wordpiece cache directory is a platform- and user-specific path where
wordpiece saves caches (such as a downloaded vocabulary). You can override
the default location in a few ways: \itemize{ \item{Option:
\code{wordpiece.dir}}{Use \code{\link{set_wordpiece_cache_dir}} to set a
specific cache directory for this session} \item{Environment:
\code{WORDPIECE_CACHE_DIR}}{Set this environment variable to specify a
wordpiece cache directory for all sessions.} \item{Environment:
\code{R_USER_CACHE_DIR}}{Set this environment variable to specify a cache
directory root for all packages that use the caching system.} }
}
